fix: remove @sanity/presentation direct dependency (#265)
This, at the same time, requires `sanity` to be downgraded in order to not run
into the `@sanity/presentation` state bug fixed in
https://github.com/sanity-io/visual-editing/pull/1496.

By removing the direct dependency we eliminate the risk of too many people
using this template with that in place. Using `@sanity/presentation` as a
direct dependency, instead of the one reexported by `sanity`, can be risky
because you need to manually keep it in line with whatever version `sanity`
supports internally. Therefore, it's better to remove it and downgrade `sanity`
and then wait a day or too until we can upgrade `sanity` and get back on the
latest version of everything.